pyproject.toml
[project]
name = "yipin-engine"
version = "0.2.0"
description = "AI-powered automated ad creative generation, placement, and fulfillment engine for Henan intangible cultural heritage food products"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.115",
    "uvicorn>=0.34",
    "anthropic>=0.52",
    "httpx>=0.28",
    "pydantic>=2.10",
    "pydantic-settings>=2.7",
    "sqlalchemy>=2.0",
    "pyyaml>=6.0",
    "streamlit>=1.42",
    "pillow>=11.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.25",
    "ruff>=0.9",
]

[tool.ruff]
line-length = 120
target-version = "py311"
